test_v1.0
This is a terst project of multiLanguage of a form or a control. The multiLanguage  surpport well in controls, but it has a problem of the Language supporting in Form. When changing the UI current cultureInfo, The categoryAttribute can not change according to the cultureInfo changing, but  properties undow the category can change.

I